# Parse floats using Parsers.jl (fixes #284)