--- a/REQUIRE
+++ b/REQUIRE
@@ -1,1 +1,0 @@
-julia 0.7-alpha