On my machine, this gives a good speedup on the (minimal) benchmarks that we have (and we now are faster than Python!):

Before:
```
$ julia1.1 bench.jl parse -f
 [BenchF] canada 0.053259659 seconds
 [BenchF] citm_catalog 0.010465748 seconds
 [BenchF] citylots 3.622005479 seconds
 [BenchF] twitter 0.004414869 seconds
 [BenchF] Total (G.M.) 0.0546397740591351 seconds
```

After:
```
$ julia1.1 bench/bench.jl parse -f
 [BenchF] canada 0.034515661 seconds
 [BenchF] citm_catalog 0.009896403 seconds
 [BenchF] citylots 3.371572698 seconds
 [BenchF] twitter 0.003937035 seconds
 [BenchF] Total (G.M.) 0.04614491419775058 seconds
```

Python:
```
$ python3 bench.py
canada 0.040348 seconds
citm_catalog 0.011868 seconds
citylots 3.509695 seconds
twitter 0.005907 seconds
Total (G.M): 0.056133
```

Cc: @quinnj @TotalVerb