@kmsquire 2019-05-30
I'll pull the transition to `Project.toml` out into its own pull request (except for the extra dependencies pulled in by `Parsers.jl`).