@kmsquire 2019-05-30
One Appveyor build failed because of a DNS lookup (unrelated).  I hesitate to close and open just to run everything again, especially if I should make any changes, and I don't have permissions to restart just that build.