@TotalVerb 2019-05-30
I'm not sure I have permission to restart appveyor either. I'm fine with the spurious failure since it's on 0.7 anyway.