@tlnagy 2019-09-10
Unfortunately, it looks like this PR was included in a JSON.jl release that still has support for 0.7 listed in its compat section in the `Project.toml` so it broke our tests over on Gadfly (actually in Compose.jl: https://travis-ci.org/GiovineItalia/Gadfly.jl/jobs/583375273). 

Here: https://github.com/JuliaRegistries/General/pull/2060/files#diff-68074356855d5796422b09f38b90317fR41

I'll open an issue over there to get this fixed retroactively.